Update README.md
If you want a solution for thunderbird backup on mac, you are going to love what you are going to find right here. There are several tools that you are going to find today which claim to backup and restore emails for you. However, most of these tools are not good enough for you.